+++ src/index.ts
@@ -2,6 +2,8 @@ import 'dotenv/config';
 
 import App from './app';
 import UserController from './user/user.controller';
+import InMemoryDatabase from './database';
+import { IUser } from './user/user.interfaces';
 
 if (!process.env.PORT) {
     process.exit(1);
@@ -9,6 +11,6 @@ if (!process.env.PORT) {
 
 const PORT: number = parseInt(process.env.PORT as string, 10);
 
-const app = new App([new UserController()], PORT);
+const app = new App([new UserController(new InMemoryDatabase<IUser>())], PORT);
 
 app.listen();
