--- src/interfaces/controller.interface.ts
@@ -1,7 +1,0 @@
-import { Router } from 'express';
-
-export interface IController {
-    path: string;
-    router: Router;
-    initializeRoutes: () => void;
-}
