+++ src/user/user.controller.ts
@@ -1,9 +1,9 @@
 import { Router, Request, Response, NextFunction } from 'express';
 import { StatusCodes } from 'http-status-codes';
 import { v4 as uuid_v4 } from 'uuid';
-import { ValidatedRequest, createValidator } from 'express-joi-validation';
+import { createValidator } from 'express-joi-validation';
 
-import { IController } from '../interfaces/controller.interface';
+import { IController } from '../interfaces/controller.interfaces';
 import InMemoryDatabase from '../database';
 import DbAdapter from '../database/db.adapter';
 import loader from '../database/loader';
@@ -11,7 +11,7 @@ import UserNotFoundException from '../exceptions/UserNotFoundException';
 import HttpException from '../exceptions/HttpException';
 import { processedDataByQueryParams } from '../utils';
 
-import { IUser } from './user.interface';
+import { IUser } from './user.interfaces';
 import * as valid from './user.validation';
 
 const PATH_TO_DUMMY_DATA = process.env.PATH_TO_DUMMY_DATA as string;
@@ -20,10 +20,11 @@ class UserController implements IController {
     public path = '/users';
     public router = Router();
 
-    private usersDb = new InMemoryDatabase<IUser>();
+    private usersDb: InMemoryDatabase<IUser>;
     private validator = createValidator();
 
-    constructor() {
+    constructor(db: InMemoryDatabase<IUser>) {
+        this.usersDb = db;
         this.initializeRoutes();
         loader(PATH_TO_DUMMY_DATA, new DbAdapter<IUser, InMemoryDatabase<IUser>>(this.usersDb));
     }
