--- src/user/user.interface.ts
@@ -1,10 +1,0 @@
-export interface IUserBase {
-    login: string;
-    password: string;
-    age: number;
-}
-
-export interface IUser extends IUserBase {
-    id: string;
-    isDeleted: boolean;
-}
